requirements.txt
# --- Web Framework & Server ---
django
channels
daphne

# --- AI & Audio Processing ---
numpy
torch # Keeping torch as requested
# REMOVED: torchaudio (very large)
# ADDED: The lightweight ONNX runtime for the VAD model
onnxruntime
silero-vad

# --- API Clients ---
mistralai==0.4.2
sarvamai
elevenlabs
python-dotenv

# --- General Utilities ---
requests
wheel
pandas